# Conflicts:
#	administrator/components/com_media/controllers/api.json.php
#	administrator/components/com_media/resources/main.js
#	media/com_media/js/mediamanager.js

+++ administrator/components/com_media/libraries/media/file/adapter/local.php
@@ -44,6 +44,44 @@ class MediaFileAdapterLocal implements MediaFileAdapterInterface
 	}
 
 	/**
+	 * Returns the requested file or folder. The returned object
+	 * has the following properties available:
+	 * - type:          The type can be file or dir
+	 * - name:          The name of the file
+	 * - path:          The relative path to the root
+	 * - extension:     The file extension
+	 * - size:          The size of the file
+	 * - create_date:   The date created
+	 * - modified_date: The date modified
+	 * - mime_type:     The mime type
+	 * - width:         The width, when available
+	 * - height:        The height, when available
+	 *
+	 * If the path doesn't exist a MediaFileAdapterFilenotfoundexception is thrown.
+	 *
+	 * @param   string  $path  The path to the file or folder
+	 *
+	 * @return  stdClass[]
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 * @throws  Exception
+	 */
+	public function getFile($path = '/')
+	{
+		// Set up the path correctly
+		$path     = JPath::clean('/' . $path);
+		$basePath = JPath::clean($this->rootPath . $path);
+
+		// Check if file exists
+		if (!file_exists($basePath))
+		{
+			throw new MediaFileAdapterFilenotfoundexception();
+		}
+
+		return $this->getPathInformation($basePath);
+	}
+
+	/**
 	 * Returns the folders and files for the given path. The returned objects
 	 * have the following properties available:
 	 * - type:          The type can be file or dir
@@ -56,6 +94,8 @@ class MediaFileAdapterLocal implements MediaFileAdapterInterface
 	 * - mime_type:     The mime type
 	 * - width:         The width, when available
 	 * - height:        The height, when available
+	 *
+	 * If the path doesn't exist a MediaFileAdapterFilenotfoundexception is thrown.
 	 *
 	 * @param   string  $path    The folder
 	 * @param   string  $filter  The filter
@@ -74,7 +114,7 @@ class MediaFileAdapterLocal implements MediaFileAdapterInterface
 		// Check if file exists
 		if (!file_exists($basePath))
 		{
-			return array();
+			throw new MediaFileAdapterFilenotfoundexception();
 		}
 
 		// Check if the path points to a file
@@ -149,6 +189,11 @@ class MediaFileAdapterLocal implements MediaFileAdapterInterface
 	 */
 	public function updateFile($name, $path, $data)
 	{
+		if (!JFile::exists($this->rootPath . $path . '/' . $name))
+		{
+			throw new MediaFileAdapterFilenotfoundexception();
+		}
+
 		JFile::write($this->rootPath . $path . '/' . $name, $data);
 	}
 
@@ -169,10 +214,20 @@ class MediaFileAdapterLocal implements MediaFileAdapterInterface
 
 		if (is_file($this->rootPath . $path))
 		{
+			if (!JFile::exists($this->rootPath . $path))
+			{
+				throw new MediaFileAdapterFilenotfoundexception();
+			}
+
 			$success = JFile::delete($this->rootPath . $path);
 		}
 		else
 		{
+			if (!JFolder::exists($this->rootPath . $path))
+			{
+				throw new MediaFileAdapterFilenotfoundexception();
+			}
+
 			$success = JFolder::delete($this->rootPath . $path);
 		}
 
@@ -196,7 +251,7 @@ class MediaFileAdapterLocal implements MediaFileAdapterInterface
 	 * - width:         The width, when available
 	 * - height:        The height, when available
 	 *
-	 * @param   string  $path    The folder
+	 * @param   string  $path  The folder
 	 *
 	 * @return  stdClass
 	 *
@@ -207,29 +262,30 @@ class MediaFileAdapterLocal implements MediaFileAdapterInterface
 		// The boolean if it is a dir
 		$isDir = is_dir($path);
 
-		// Set the values
-		$obj                = new stdClass;
-		$obj->type          = $isDir ? 'dir' : 'file';
-		$obj->name          = basename($path);
-		$obj->path          = str_replace($this->rootPath, '/', $path);
-		$obj->extension     = !$isDir ? JFile::getExt($obj->name) : '';
-		$obj->size          = !$isDir ? filesize($path) : 0;
-		$obj->create_date   = $this->getDate(filectime($path))->format('c', true);
-		$obj->modified_date = $this->getDate(filemtime($path))->format('c', true);
-		$obj->mime_type     = mime_content_type($path);
+		$createDate   = $this->getDate(filectime($path));
+		$modifiedDate = $this->getDate(filemtime($path));
 
-		try
+		// Set the values
+		$obj                          = new stdClass;
+		$obj->type                    = $isDir ? 'dir' : 'file';
+		$obj->name                    = basename($path);
+		$obj->path                    = str_replace($this->rootPath, '/', $path);
+		$obj->extension               = !$isDir ? JFile::getExt($obj->name) : '';
+		$obj->size                    = !$isDir ? filesize($path) : 0;
+		$obj->create_date             = $createDate->format('c', true);
+		$obj->create_date_formatted   = (string) $createDate; // TODO use format from config
+		$obj->modified_date           = $modifiedDate->format('c', true);
+		$obj->modified_date_formatted = (string) $modifiedDate; // TODO use format from config
+		$obj->mime_type               = mime_content_type($path);
+		$obj->width                   = 0;
+		$obj->height                  = 0;
+
+		if (strpos($obj->mime_type, 'image/') === 0 && in_array(strtolower($obj->extension), array('jpg', 'jpeg', 'png', 'gif', 'bmp')))
 		{
 			// Get the image properties
 			$props       = JImage::getImageFileProperties($path);
 			$obj->width  = $props->width;
 			$obj->height = $props->height;
-		}
-		catch (Exception $e)
-		{
-			// Probably not an image
-			$obj->width  = 0;
-			$obj->height = 0;
 		}
 
 		return $obj;
@@ -249,7 +305,8 @@ class MediaFileAdapterLocal implements MediaFileAdapterInterface
 		$dateObj = JFactory::getDate($date);
 
 		$timezone = JFactory::getApplication()->get('offset');
-		$user = JFactory::getUser();
+		$user     = JFactory::getUser();
+
 		if ($user->id)
 		{
 			$userTimezone = $user->getParam('timezone');
